match uchar type in isvalidcube()

+++ src/engine/world/octacube.cpp
@@ -132,10 +132,10 @@ bool cube::isvalidcube() const
     genclipbounds(*this, ivec(0, 0, 0), 256, p);
     genclipplanes(*this, ivec(0, 0, 0), 256, p);
     // test that cube is convex
-    for(uint i = 0; i < 8; ++i)
+    for(uchar i = 0; i < 8; ++i)
     {
         vec v = p.v[i];
-        for(uint j = 0; j < p.size; ++j)
+        for(uchar j = 0; j < p.size; ++j)
         {
             if(p.p[j].dist(v)>1e-3f)
             {
